function/requirements.txt
azure-functions==1.6.0
psycopg2-binary==2.8.6
sendgrid